requirements.txt
fastapi
uvicorn
python-multipart
passlib==1.7.4
sqlalchemy
pyjwt
python-dotenv
bcrypt>=4.0.1
psutil
pydantic[email]
email-validator
PyPDF2
python-docx
chardet